+++ lib/ui/main_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import 'database/database_page.dart';
 import 'kinds/kinds_page.dart';
 import 'main_screen_providers.dart';
 import 'products/products_page.dart';
@@ -13,6 +14,7 @@ import 'widgets/weekly_overview_panel.dart';
 /// - Products section
 /// - Kinds section
 /// - Recipes section
+/// - Database section
 class MainScreen extends ConsumerWidget {
   const MainScreen({super.key});
 
@@ -33,6 +35,8 @@ class MainScreen extends ConsumerWidget {
         return const KindsPage();
       case AppSection.recipes:
         return const RecipesPage();
+      case AppSection.database:
+        return const DatabasePage();
     }
   }
 }